Implemented a better collision system. Also, collision bug fixed.

+++ core/src/main/java/knight/arkham/objects/Brick.java
@@ -2,39 +2,41 @@ package knight.arkham.objects;
 
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.math.Rectangle;
+import com.badlogic.gdx.physics.box2d.Body;
 import com.badlogic.gdx.physics.box2d.World;
 import knight.arkham.helpers.Box2DBody;
-import knight.arkham.helpers.ContactType;
+import knight.arkham.helpers.Box2DHelper;
 
 public class Brick extends GameObject{
-
-    private final World world;
     private boolean isDestroyed;
     private boolean setToDestroy;
-    public Brick(Rectangle rectangle, World world, String spritePath) {
-        super(
-            new Box2DBody(rectangle, 0, world, ContactType.BRICK),
-            spritePath, 0
+    public Brick(Rectangle bounds, World world, String spritePath) {
+        super(bounds, world, spritePath, 0);
+    }
+
+    @Override
+    protected Body createBody() {
+        return Box2DHelper.createBody(
+            new Box2DBody(actualBounds, 0, actualWorld, this)
         );
-        this.world = world;
     }
 
     public void update(){
 
-        if (setToDestroy)
+        if (setToDestroy && !isDestroyed)
             destroyBrick();
+    }
+
+    private void destroyBrick() {
+
+        actualWorld.destroyBody(body);
+        isDestroyed = true;
     }
 
     public void draw(Batch batch){
 
         if (!isDestroyed)
             super.draw(batch);
-    }
-
-    private void destroyBrick() {
-
-        world.destroyBody(body);
-        isDestroyed = true;
     }
 
     public void hitByBall() {
